Move Trackable, Tag interface under to Legacy.ts
