+++ src/trackers/ecommerce/index.ts
@@ -1,2 +1,3 @@
 export * from './common';
 export * from './ga';
+export * from './legacy';
